Updated the text which is displayed when the user hovers over each of the figures
Action for Melissa: Would be great if you could double-check all the inputs (i.e. the correct value and column value displayed for each label).

Please ignore the changes at the beginning to do with the loading of packages, not sure what was going wrong so just found something that worked as a workaround and did not spend lots of time investigating the issue. My session also kept crashing which I assumed had to do with the same issue